requirements.txt
# Prompt Injection Detector - Requirements
# Compatible with macOS (Apple Silicon M4 Pro)
# Python 3.10+

# Core ML Libraries
numpy>=1.26.0,<2.0.0
pandas>=2.2.0
scikit-learn>=1.4.0
scipy>=1.12.0

# Gradient Boosting
xgboost>=2.0.0
lightgbm>=4.3.0

# Deep Learning & Transformers
torch>=2.2.0
transformers>=4.40.0
sentence-transformers>=3.0.0

# NLP Libraries
nltk>=3.8.1
spacy>=3.7.0

# Text Analysis
py-readability-metrics>=1.4.5
textstat>=0.7.4
pyspellchecker>=0.7.3

# Utilities
tqdm>=4.66.0
joblib>=1.3.0
python-dotenv>=1.0.0
pyyaml>=6.0

# API/Service (for deployment)
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# Model Export
onnx>=1.15.0
skl2onnx>=1.16.0
onnxruntime>=1.17.0

# Data Handling
datasets>=2.17.0
huggingface-hub>=0.20.0

# Development/Testing (optional - install separately if needed)
# matplotlib>=3.8.0  # REMOVED - causes build issues on macOS
# seaborn>=0.13.0    # REMOVED - depends on matplotlib
# pytest>=8.0.0
# pytest-cov>=4.1.0
# black>=24.0.0
